Move class initialization and object instantiation tracing into static state
