+++ sdk/src/org.graalvm.nativeimage/src/org/graalvm/nativeimage/impl/RuntimeClassInitializationSupport.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -53,8 +53,4 @@ public interface RuntimeClassInitializationSupport {
     void rerunInitialization(Class<?> aClass, String reason);
 
     void initializeAtBuildTime(Class<?> aClass, String reason);
-
-    void reportClassInitialized(Class<?> aClass, StackTraceElement[] stackTrace);
-
-    void reportObjectInstantiated(Object o, StackTraceElement[] stackTrace);
 }
