+++ sdk/src/org.graalvm.nativeimage/src/org/graalvm/nativeimage/impl/clinit/ClassInitializationTracking.java
@@ -44,6 +44,13 @@ import org.graalvm.nativeimage.ImageSingletons;
 import org.graalvm.nativeimage.impl.ImageSingletonsSupport;
 import org.graalvm.nativeimage.impl.RuntimeClassInitializationSupport;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.IdentityHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
 public class ClassInitializationTracking {
 
     /**
@@ -55,6 +62,14 @@ public class ClassInitializationTracking {
      */
     public static final boolean IS_IMAGE_BUILD_TIME;
 
+    public static final Map<Class<?>, StackTraceElement[]> initializedClasses = new ConcurrentHashMap<>();
+    /**
+     * Instantiated objects must be traced using their identities as their hashCode may change
+     * during the execution. We also want two objects of the same class that have the same hash and
+     * are equal to be mapped as two distinct entries.
+     */
+    public static final Map<Object, StackTraceElement[]> instantiatedObjects = Collections.synchronizedMap(new IdentityHashMap<>());
+
     static {
         /*
          * Prevents javac from constant folding use of this field. It is set to true by the process
@@ -64,25 +79,46 @@ public class ClassInitializationTracking {
     }
 
     /**
-     * This method is called from the instrumented class initialization methods.
+     * This method is called from the native-image diagnostics agent to report class initialization.
      */
     @SuppressWarnings({"unused", "ConstantConditions"})
-    public static void reportClassInitialized(Class<?> c, StackTraceElement[] stackTrace) {
-        if (ImageSingletonsSupport.isInstalled() && ImageSingletons.contains(RuntimeClassInitializationSupport.class)) {
-            RuntimeClassInitializationSupport runtimeClassInitialization = ImageSingletons.lookup(RuntimeClassInitializationSupport.class);
-            runtimeClassInitialization.reportClassInitialized(c, stackTrace);
-        }
+    public static void reportClassInitialized(Class<?> clazz, StackTraceElement[] stackTrace) {
+        initializedClasses.put(clazz, relevantStackTrace(stackTrace));
     }
 
     /**
-     * This method is called from the instrumented class initialization methods.
+     * This method is called from the native-image-diagnostics agent to report object instantiation.
      */
     @SuppressWarnings({"unused"})
     public static void reportObjectInstantiated(Object o, StackTraceElement[] stackTrace) {
-        if (ImageSingletonsSupport.isInstalled() && ImageSingletons.contains(RuntimeClassInitializationSupport.class)) {
-            RuntimeClassInitializationSupport runtimeClassInitialization = ImageSingletons.lookup(RuntimeClassInitializationSupport.class);
-            runtimeClassInitialization.reportObjectInstantiated(o, stackTrace);
+        instantiatedObjects.putIfAbsent(o, relevantStackTrace(stackTrace));
+    }
+
+    /**
+     * If the stack trace contains class initializaiton takes the stack up to the last
+     * initialization. Otherwise returns the whole stack trace. The method never returns the stack
+     * from the instrumented part.
+     *
+     * This method can be refined on a case-by-case basis to print nicer traces.
+     *
+     * @return a stack trace that led to erroneous situation
+     */
+    private static StackTraceElement[] relevantStackTrace(StackTraceElement[] stack) {
+        ArrayList<StackTraceElement> filteredStack = new ArrayList<>();
+        int lastClinit = 0;
+        boolean containsLambdaMetaFactory = false;
+        for (int i = 0; i < stack.length; i++) {
+            StackTraceElement stackTraceElement = stack[i];
+            if ("<clinit>".equals(stackTraceElement.getMethodName())) {
+                lastClinit = i;
+            }
+            if (stackTraceElement.getClassName().equals("java.lang.invoke.LambdaMetafactory")) {
+                containsLambdaMetaFactory = true;
+            }
+            filteredStack.add(stackTraceElement);
         }
+        List<StackTraceElement> finalStack = lastClinit != 0 && !containsLambdaMetaFactory ? filteredStack.subList(0, lastClinit + 1) : filteredStack;
+        return finalStack.toArray(new StackTraceElement[0]);
     }
 
 }
